Move views tests for comments to test_views
